+++ nvim/init.lua
@@ -29,6 +29,10 @@ require('lazy').setup({
   -- Kickstart plugins (neo-tree, autopairs, gitsigns, indent guides)
   { import = 'kickstart.plugins' },
 }, {
+  dev = {
+    path = '~/playground',
+    fallback = true,
+  },
   ui = {
     icons = vim.g.have_nerd_font and {} or {
       cmd = '⌘',